requirements.txt
# Browser-Use API Server Dependencies
# Most of these are already installed in the venv except aiosqlite and python-multipart

browser-use>=0.11.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-multipart
